Don't use CANVAS_ORIGIN anymore
The canvas origin existed because parts positioned at (512,512) appeared to be in the center in-game. The offset of 512 is added in the file parsing routine in the game, so this is moved to our (de)serialization processing and the CANVAS_ORIGIN macro is reset to zero. It will be removed later